refactor(components): refactor backtop (#3486)
